cleanup config/config.go, remove obsolete wallet connection code
