+++ main.go
@@ -5,14 +5,11 @@ import (
 	"fmt"
 	"os"
 	"os/signal"
-	"path/filepath"
-	"strings"
 	"sync"
 	"syscall"
 
 	"github.com/raedahgroup/godcr/app"
 	"github.com/raedahgroup/godcr/app/config"
-	"github.com/raedahgroup/godcr/app/walletmediums/dcrlibwallet"
 	"github.com/raedahgroup/godcr/fyne"
 )
 
@@ -26,7 +23,8 @@ var shutdownOps []func()
 var opError error
 
 func main() {
-	appConfig, args, err := config.LoadConfig()
+	// nb: cli support will require loading from a config file
+	cfg, err := config.LoadConfigFromDb()
 	if err != nil {
 		fmt.Fprintln(os.Stderr, err.Error())
 		os.Exit(1)
@@ -42,23 +40,17 @@ func main() {
 	}()
 
 	// Special show command to list supported subsystems and exit.
-	if appConfig.DebugLevel == "show" {
+	if cfg.DebugLevel == "show" {
 		fmt.Println("Supported subsystems", supportedSubsystems())
 		os.Exit(0)
 	}
 
 	// Parse, validate, and set debug log level(s).
-	if err := parseAndSetDebugLevels(appConfig.DebugLevel); err != nil {
+	if err := parseAndSetDebugLevels(cfg.DebugLevel); err != nil {
 		err := fmt.Errorf("loadConfig: %s", err.Error())
 		fmt.Fprintln(os.Stderr, err)
 		os.Exit(1)
 		return
-	}
-
-	// check if user passed commands/options/args but is not running in cli mode
-	if appConfig.InterfaceMode != "cli" && len(args) > 0 {
-		fmt.Fprintf(os.Stderr, "unexpected command or flag in %s mode: %s\n", appConfig.InterfaceMode, strings.Join(args, " "))
-		os.Exit(1)
 	}
 
 	// use wait group to keep main alive until shutdown completes
@@ -73,7 +65,7 @@ func main() {
 	shutdownOps = append(shutdownOps, cancel)
 
 	// open connection to wallet and add wallet close function to shutdownOps
-	walletMiddleware, err := connectToWallet(ctx, appConfig)
+	walletMiddleware, err := connectToWallet(ctx, cfg)
 	if err != nil {
 		fmt.Fprintln(os.Stderr, "Failed to connect to wallet.", err.Error())
 		fmt.Println("Exiting.")
@@ -109,35 +101,7 @@ func logWarn(message string) {
 // default is connecting directly to a wallet database file via dcrlibwallet
 // alternative is connecting to wallet database via dcrwallet rpc (if rpc server address is provided)
 func connectToWallet(ctx context.Context, cfg *config.Config) (app.WalletMiddleware, error) {
-	return connectViaDcrlibwallet(ctx, cfg)
-}
-
-// connectViaDcrWalletRPC attempts to load the database at `cfg.DefaultWalletDir`.
-// Prompts user to select wallet to connect to if default wallet dir isn't set
-// or wallet could not be found at set default dir.
-func connectViaDcrlibwallet(ctx context.Context, cfg *config.Config) (*dcrlibwallet.DcrWalletLib, error) {
-	// attempt to load default wallet if set and wallet db can be found
-	if cfg.DefaultWalletDir != "" {
-		netType := filepath.Base(cfg.DefaultWalletDir)
-		walletMiddleware, err := dcrlibwallet.Connect(ctx, cfg.DefaultWalletDir, netType)
-		if err != nil {
-			return nil, err
-		}
-
-		defaultWalletExists, walletCheckError := walletMiddleware.WalletExists()
-		if walletCheckError != nil {
-			return nil, fmt.Errorf("\nError checking default wallet directory for wallet database.\n%s",
-				walletCheckError.Error())
-		}
-
-		if defaultWalletExists {
-			fmt.Println("Using wallet", cfg.DefaultWalletDir)
-			return walletMiddleware, nil
-		}
-	}
-
-	// todo: no longer detecting wallets, so should not require default dir
-	return nil, fmt.Errorf("default wallet not configured")
+	return nil, fmt.Errorf("wallet connection middleware not yet implemented")
 }
 
 func enterFyneMode(ctx context.Context, walletMiddleware app.WalletMiddleware) {
